moved download scripts to its own class
