+++ bin/index.js
@@ -13,3 +13,5 @@ module.exports.frames = require('./actions/frames');
 module.exports.keys = require('./actions/keys');
 
 module.exports.Uploader = require('./actions/uploader');
+
+module.exports.Downloader = require('./actions/downloader');
